move workaround out of sorting models file
